Moved classes to new directories.
