+++ src/Utils/Container.php
@@ -12,11 +12,11 @@
  * @link https://github.com/jaxon-php/jaxon-core
  */
 
-namespace Jaxon;
+namespace Jaxon\Utils;
 
 use Lemon\Event\EventDispatcher;
 
-use Jaxon\App;
+use Jaxon\App\App;
 use Jaxon\Response\Response;
 use Jaxon\Config\Config;
 use Jaxon\Config\Reader as ConfigReader;
@@ -31,17 +31,17 @@ use Jaxon\Request\Factory as RequestFactory;
 use Jaxon\Response\Manager as ResponseManager;
 use Jaxon\Plugin\Manager as PluginManager;
 use Jaxon\Plugin\CodeGenerator;
-use Jaxon\Plugin\Bootstrap;
-use Jaxon\Ui\View\Manager as ViewManager;
-use Jaxon\Ui\View\Facade as ViewFacade;
-use Jaxon\Ui\View\Renderer;
-use Jaxon\Ui\Dialogs\Dialog;
-use Jaxon\Ui\Template\Minifier;
-use Jaxon\Ui\Template\Engine as TemplateEngine;
-use Jaxon\Ui\Pagination\Paginator;
-use Jaxon\Ui\Pagination\Renderer as PaginationRenderer;
-use Jaxon\Str\Validation\Validator;
-use Jaxon\Str\Translation\Translator;
+use Jaxon\App\Bootstrap;
+use Jaxon\Utils\View\Manager as ViewManager;
+use Jaxon\Utils\View\Facade as ViewFacade;
+use Jaxon\Utils\View\Renderer;
+use Jaxon\Utils\Dialogs\Dialog;
+use Jaxon\Utils\Template\Minifier;
+use Jaxon\Utils\Template\Engine as TemplateEngine;
+use Jaxon\Utils\Pagination\Paginator;
+use Jaxon\Utils\Pagination\Renderer as PaginationRenderer;
+use Jaxon\Utils\Validation\Validator;
+use Jaxon\Utils\Translation\Translator;
 use Jaxon\Contracts\Template\Renderer as TemplateRenderer;
 use Jaxon\Contracts\Session as SessionContract;
 use Jaxon\Contracts\Container as ContainerContract;
@@ -69,8 +69,8 @@ class Container
     {
         $this->libContainer = new \Pimple\Container();
 
-        $sTranslationDir = realpath(__DIR__ . '/../translations');
-        $sTemplateDir = realpath(__DIR__ . '/../templates');
+        $sTranslationDir = realpath(__DIR__ . '/../../translations');
+        $sTemplateDir = realpath(__DIR__ . '/../../templates');
         $this->init($sTranslationDir, $sTemplateDir);
     }
