+++ src/Utils/Pagination/Renderer.php
@@ -12,7 +12,7 @@
  * @link https://github.com/jaxon-php/jaxon-core
  */
 
-namespace Jaxon\Ui\Pagination;
+namespace Jaxon\Utils\Pagination;
 
 use Jaxon\Contracts\Template\Renderer as TemplateRenderer;
